Refactor client
* Code is kept in a separate file
* Support for basic auth via flags
* Uses environment variables to read some parameters
